Input to sBoom and output contour
Created sBoom_input.py which creates a dictionary in a form that should be easy to use in sBoom. Created noise_contour which needs the output data from pyLdB but then will make the noise contour over the U.S. Also, new data files are included and edits to functions.py were made.